+++ pointcloudset/io/__init__.py
@@ -6,9 +6,9 @@ from pyntcloud.io import (
     FROM_FILE as POINTCLOUD_FROM_FILE,
 )  # needs to be here, ignore warnings
 
-from pointcloudset.io.dataset.ros import dataset_from_ros
 from pointcloudset.io.dataset.dir import dataset_from_dir, dataset_to_dir
 from pointcloudset.io.dataset.pointcloud import dataset_from_pointclouds
+from pointcloudset.io.dataset.ros import dataset_from_ros
 from pointcloudset.io.pointcloud.csv import write_csv
 from pointcloudset.io.pointcloud.open3d import from_open3d, to_open3d
 from pointcloudset.io.pointcloud.pandas import from_dataframe, to_dataframe